Move #18 : data class들 따로 빼서 data 패키지에 이동
