+++ app/src/main/java/org/sopt/cgv/feature/time/component/SelectableTheatersInModal.kt
@@ -34,12 +34,7 @@ import org.sopt.cgv.core.designsystem.theme.Gray700
 import org.sopt.cgv.core.designsystem.theme.Gray850
 import org.sopt.cgv.core.designsystem.theme.PrimaryRed400
 import org.sopt.cgv.core.designsystem.theme.White
-
-
-data class MovieTheatersByDetailRegion(
-    val detailRegionName: String,
-    val theaterNames: PersistentList<String>
-)
+import org.sopt.cgv.feature.time.data.MovieTheatersByDetailRegion
 
 @Composable
 fun SelectableTheatersInModal(
