+++ app/src/main/java/org/sopt/cgv/feature/time/component/TimeScreenDateSelectionTab.kt
@@ -30,6 +30,7 @@ import org.sopt.cgv.core.designsystem.theme.CGVTheme
 import org.sopt.cgv.core.designsystem.theme.Gray700
 import org.sopt.cgv.core.designsystem.theme.PrimaryRed400
 import org.sopt.cgv.core.designsystem.theme.White
+import org.sopt.cgv.feature.time.data.Date
 
 @Composable
 fun TimeScreenDateSelectionTab(
@@ -174,8 +175,3 @@ private fun TimeScreenDateSelectionTabPreview() {
     )
 
 }
-
-data class Date(
-    val date: String,
-    val day: String
-)
