+++ app/src/main/java/org/sopt/cgv/feature/time/data/Date.kt
@@ -1,0 +1,6 @@
+package org.sopt.cgv.feature.time.data
+
+data class Date(
+    val date: String,
+    val day: String
+)
