+++ app/src/main/java/org/sopt/cgv/feature/time/data/MovieTheatersByDetailRegion.kt
@@ -1,0 +1,8 @@
+package org.sopt.cgv.feature.time.data
+
+import kotlinx.collections.immutable.PersistentList
+
+data class MovieTheatersByDetailRegion(
+    val detailRegionName: String,
+    val theaterNames: PersistentList<String>
+)
